+++ packages/time-aware-polyline/CHANGELOG.md
@@ -2,3 +2,4 @@
 - Ported `time-aware-polyline-js` in typescript
 - Add polyline utility class
 - Add timeAwarePolyline Utility class
+- Init/Update animation from encoded polyline or time aware array [lat, lng, timestamp]